Do not set base class variables in tests
Each test class was setting class variables in the base class
directly. This meant that if one test class did not reset everything
the tests would not work depending on which tests had already run.

This commit changes the behvior so that the test values are only
set in the instance object rather than globally in the base class.
Config classes are now instance local as well.