RingsSkyMap: fix findAllTracts for poles
The initial pole check was wrong (doesn't allow other tracts to be
included) and unnecessary because the poles get picked up by the
check at the end.

Added a test.

Thought about allowing negative indexing into SkyMap, but decided
that would be a mistake because the SkyMap isn't really like a list,
but more like a dict.