Several updates to examples/showVisitSkyMap.py
These include making a color map such that the outline for each visit
gets a unique color and the addition of a legend with the visit numbers
such that they can be identified on the plot.  Also adds the ability to
specify multiple tracts in the input and their outlines are drawn on the
plot with unique gray-scaling and added to the legend.

This also fixes an issue with using a skymap with multiple tracts.  The
previous code misleadingly fixed the tract number to 0, but still iterated
over all patches in all tracts in the given skymap (as opposed to just the
one specified in the --id), leading to extremely long run-times for full
sky maps.