RingsSkyMap: fix bugs in tract numbering
getRingIndices, findTract and findTractAll had off-by-one bugs.
These affect the setting of primary flags (e.g. detect_isPrimary) in
pipe_tasks for tracts near the RA=360 line. However, this involves
renumbering some of the tracts (specifically, those on the RA=360 line).

We use the existing version mechanism to support the old tract numbering.
version=0 is valid up until DM-14809, version=1 is valid afterwards.
The version is preserved in the skymap pickle, so this Just Works.
For version=0, we leave the tract numbering as it is (broken in two places:
there is one duplicated tract (the first tract on the first ring) and
one missing tract (the first tract on the last ring)), but fix the lookup
in the findTract method to work with the broken numbering.

Included in these changes is a fix for non-zero raStart: simply using fmod
doesn't work to wrap angles because it leaves negative values as negative.
Using Angle.wrap works.

Added specific tests for RingsSkyMap and generic tests for all skymaps
to catch these sorts of bugs.

Thanks to Sogo Mineo (NAOJ) for finding and diagnosing these bugs!