BaseSkyMap.register: optimise memory usage
Registering the HSC rings skymap is triggering the OOM-killer
on my Docker with 8 GB memory. The main problem is constructing
the records before they are needed, so all records are in memory
at once. Instead, create the records in a loop within the
transaction context.

This greatly reduces the memory footprint because we don't need
all records in memory at once. This comes at the price of
repeated calls to registry.insertDimensionData, but even if
there is a small loss of performance, this is a command that
gets run only once at the setting up of a Gen3 datastore, so
it shouldn't be too burdensome.