Add ability to impose a miminum overlap fraction
Also do some limit setting gymnastics to try to ensure that the
aspect ratio of the plot is in focal plane coordinates (i.e. such
that a square detector appears as a square).  The conventions of
HSC make this somewhat finicky, but all cases tested (including
LATISS, DC2, and DECam datasets in addition to HSC collections that
include rotations) come out as expected.